+++ ft_strnew_two.c
@@ -6,13 +6,13 @@
 /*   By: cfatrane <cfatrane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/20 11:36:53 by cfatrane          #+#    #+#             */
-/*   Updated: 2017/01/20 14:10:19 by cfatrane         ###   ########.fr       */
+/*   Updated: 2017/01/27 20:12:54 by cfatrane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	**ft_strnew_two(int x, int y)
+char	**ft_strnew_two(size_t x, size_t y)
 {
 	int		i;
 	char	**tab;
